Move check_if_multimodal
